fix: eliminate startup race, fix routing receiver, add watchdog timer
Root causes fixed:
- CRITICAL: startChaosMode now passes preset+boost directly in the FGS
  intent instead of a separate broadcast, eliminating the race condition
  where START_CHAOS_MODE broadcast fired before commandReceiver was
  registered in onCreate() -- this caused the service to never reach
  ACTIVE state, keeping the UI stuck at 'Starting...' forever.

- CRITICAL: routingReceiver is now registered AFTER serviceState is set
  to ACTIVE (was registered in onCreate), preventing the sticky
  ACTION_HEADSET_PLUG broadcast from false-triggering a session teardown
  at startup.

- CRITICAL: battery optimization exemption dialog moved off the critical
  startup path -- now requested asynchronously after ACTIVE is confirmed.

- Added 12-second startup watchdog timer in ChaosController: if the
  native service never emits ACTIVE within the window, auto-fail with a
  clear error message and reset to stopped state.

- Simplified startup from 4 steps to 3: notification permission +
  mic permission + startChaosMode (single native call).

- User can now tap the button during 'Starting...' to cancel a stuck
  loading state (removed the isLoading early-return guard).

- Fixed Dart type errors in effect_engine.dart (num vs double in
  saturation stages).

- Removed unused imports (dart:io from main.dart, volume_boost_widget
  from home_screen.dart).

+++ lib/effect_engine.dart
@@ -131,8 +131,8 @@ class EffectEngine {
     }
 
     // ── Stage 3: Saturation 1 ────────────────────────────────────────────
-    final g1 = saturation1Gain.clamp(0.1, 50);
-    final t1 = _tanh(g1).abs().clamp(1e-9, double.infinity);
+    final g1 = saturation1Gain.clamp(0.1, 50).toDouble();
+    final t1 = _tanh(g1).abs().clamp(1e-9, double.infinity).toDouble();
     for (var i = 0; i < len; i++) work[i] = _tanh(g1 * work[i]) / t1;
 
     // ── Stage 4: Feedback Delay ──────────────────────────────────────────
@@ -148,8 +148,8 @@ class EffectEngine {
     }
 
     // ── Stage 5: Saturation 2 ────────────────────────────────────────────
-    final g2 = saturation2Gain.clamp(0.1, 20);
-    final t2 = _tanh(g2).abs().clamp(1e-9, double.infinity);
+    final g2 = saturation2Gain.clamp(0.1, 20).toDouble();
+    final t2 = _tanh(g2).abs().clamp(1e-9, double.infinity).toDouble();
     for (var i = 0; i < len; i++) work[i] = _tanh(g2 * work[i]) / t2;
 
     // ── Stage 6: Ring Modulation ─────────────────────────────────────────
